#60 move entities to separate module
