+++ backend/src/models/entity/plants_impl.rs
@@ -1,0 +1,21 @@
+//! Contains the implementation of [`Plants`].
+
+use diesel::{PgConnection, QueryDsl, QueryResult, RunQueryDsl};
+
+use crate::{
+    models::dto::plants::PlantsDTO,
+    schema::plants::{self, all_columns},
+};
+
+use super::Plants;
+
+impl Plants {
+    /// Fetch all plants from the database.
+    ///
+    /// # Errors
+    /// * Unknown, diesel doesn't say why it might error.
+    pub fn find_all(conn: &mut PgConnection) -> QueryResult<Vec<PlantsDTO>> {
+        let query_result = plants::table.select(all_columns).load::<Self>(conn);
+        query_result.map(|v| v.into_iter().map(Into::into).collect())
+    }
+}
