+++ backend/src/models/mod.rs
@@ -1,7 +1,6 @@
 //! Entities and DTOs used in `PermaplanT`.
 
+pub mod entity;
 pub mod dto;
 pub mod r#enum;
-pub mod plants;
 pub mod response;
-pub mod seed;
